MDL-37122 core get_max_upload_sizes: Include "Course/Site/Activity limit (X)" in list of options
This change removes the "0 bytes" option from the get_max_upload_sizes list
and replaces it with "Course limit (X)" or "Site limit (X)" (whichever is smaller).
This means we can remove all custom handling in the modules that were removing and
adding these options. It only affects pages that pass valid options for sitelimit and
courselimit - so admin pages will work correctly.

It also orders the list so the course/site limit options will be first
(as it will be the largest).

AMOS START
 REM [courseuploadlimit, core] has been parameterized to get [uploadlimitwithsize, core]
AMOS END